Put service-param-standardization in own file
