+++ backend/api/routes/service.js
@@ -10,11 +10,13 @@ import {
 import {
   checkServiceName,
   checkServiceId,
+} from "#utils/api-helpers/service-utils.js";
+import {
   validateAndFormatServiceData,
   prettyPrintServicesOfferedList,
   reformatServiceForReturn,
   dateString,
-} from "#utils/api-helpers/service-utils.js";
+} from "#utils/api-helpers/service-param-standardize.js";
 import { createErrorReturn } from "#utils/validate-utils.js";
 import searchServices from "#search/search-services.js";
 import getFilter from "#utils/search-filters/filter-create-utils.js";
