+++ backend/utils/recs/rank-by-keyword.js
@@ -1,6 +1,6 @@
 // Local Imports
 import serviceTypeKeywords from "#recs/serviceTypeKeywords.json" with { type: "json" };
-import { prettyPrintService } from "#utils/api-helpers/service-utils.js";
+import { prettyPrintService } from "#utils/api-helpers/service-param-standardize.js";
 
 /**
  * Rank services based how many serviceTypes they offer based on keyword matches
